requirements.txt
Flask>=1.1.1
PyGithub>=1.43.8
requests>=2.22.0
pydantic

